Use agreed-typed
