--- spec/agreed/agreed.js
@@ -1,13 +1,0 @@
-module.exports = [
-  "./agreedsample/failure_get.js",
-  "./agreedsample/get.js",
-  "./uploadsample/post.js",
-  "./search/getSearchById.js",
-  "./search/getSearchList.js",
-  "./style/getStyle.js",
-  "./masters/getGenderMaster.js",
-  "./masters/getHairLengthMaster.js",
-  "./masters/getHairColorMaster.js",
-  "./masters/getAreaMaster.js",
-  "./masters/getMenuContentsMaster.js",
-];
