+++ spec/agreed/agreed.ts
@@ -1,0 +1,17 @@
+import { convert } from "agreed-typed";
+import { flatten } from "lodash";
+
+module.exports = convert(
+  ...flatten([
+    require("./agreedsample/get"),
+    require("./uploadsample/post"),
+    require("./search/getSearchById"),
+    require("./search/getSearchList"),
+    require("./style/getStyle"),
+    require("./masters/getGenderMaster"),
+    require("./masters/getHairLengthMaster"),
+    require("./masters/getHairColorMaster"),
+    require("./masters/getAreaMaster"),
+    require("./masters/getMenuContentsMaster"),
+  ]),
+);
