--- spec/agreed/agreedsample/failure_get.js
@@ -1,20 +1,0 @@
-module.exports = {
-  request: {
-    path: "/agreedsample",
-    method: "GET",
-    query: {
-      status: "{:status}",
-    },
-  },
-  response: {
-    status: "{:status}",
-    body: {
-      results: {
-        text: "{:text}",
-      },
-    },
-    values: {
-      text: "Hello world",
-    },
-  },
-};
