--- spec/agreed/agreedsample/get.js
@@ -1,16 +1,0 @@
-module.exports = {
-  request: {
-    path: "/agreedsample",
-    method: "GET",
-  },
-  response: {
-    body: {
-      results: {
-        text: "{:text}",
-      },
-    },
-    values: {
-      text: "Hello world",
-    },
-  },
-};
