+++ spec/agreed/search/getSearchById.ts
@@ -1,0 +1,29 @@
+import { APIDef, GET, Success200, ResponseDef, Error404 } from "agreed-typed";
+module.exports = [
+  {
+    request: {
+      path: "/beauty/search/:id",
+      method: "GET",
+    },
+    response: {
+      body: {
+        results: {
+          search: [
+            {
+              name: "{:name}",
+              urls: {
+                pc: "{:pc}",
+                mobile: "{:mobile}",
+              },
+            },
+          ],
+        },
+      },
+      values: {
+        name: "サロン",
+        pc: "/public/cat_large.jpeg",
+        mobile: "/public/cat_small.jpeg",
+      },
+    },
+  },
+];
