--- spec/agreed/uploadsample/post.js
@@ -1,15 +1,0 @@
-module.exports = {
-  request: {
-    path: "/uploadsample",
-    method: "POST",
-    body: {
-      path: "{:path}",
-    },
-    values: {
-      path: "/public/hoge",
-    },
-  },
-  response: {
-    status: 201,
-  },
-};
