+++ src/server/services/__tests__/AgreedSample.read.test.ts
@@ -4,7 +4,8 @@ import configs from "../../configs";
 import AgreedSample from "../AgreedSample";
 
 test("AgreedSample: read success", async () => {
-  const getText = require("../../../../spec/agreed/agreedsample/get.js");
+  const getText = require("../../../../spec/agreed/agreedsample/get");
+
   const agreedSample = new AgreedSample(configs);
   const result = await (agreedSample.read as any)({}, {}, {});
   assert.deepStrictEqual(result, getText.response.values);
