+++ src/server/services/__tests__/AgreedService.read.test.ts
@@ -4,22 +4,13 @@ import config from "../../configs/test";
 import AgreedService from "../AgreedService";
 
 test("AgreedService: read success", async () => {
-  const getText = require("../../../../spec/agreed/agreedsample/get.js");
+  const getText = require("../../../../spec/agreed/agreedsample/get.ts");
   const agreedService: any = new AgreedService(
     config,
     "foo",
     "/agreedsample",
     "foo",
   );
-  const result = await agreedService.read();
-  assert.deepStrictEqual(result, getText.response.values);
-});
-
-test("AgreedService: read failure", async () => {
-  const getText = require("../../../../spec/agreed/agreedsample/failure_get.js");
-  const agreedService: any = new AgreedService(config, "foo", "/agreedsample", {
-    stats: 404,
-  });
   const result = await agreedService.read();
   assert.deepStrictEqual(result, getText.response.values);
 });
