+++ tsconfig.json
@@ -58,6 +58,6 @@
     // "experimentalDecorators": true,        /* Enables experimental support for ES7 decorators. */
     // "emitDecoratorMetadata": true,         /* Enables experimental support for emitting type metadata for decorators. */
   },
-  "include": ["src/**/*.ts"],
+  "include": ["src/**/*", "spec/**/*"],
   "exclude": ["node_modules"]
 }
